Updates quiz summary page and how to play widget.
Updates the quiz summary page to increase the time for viewing results to 30 seconds and the time for the next quiz to 5 seconds. Removes the text displaying the time for viewing results and updates the text for the next quiz. Modifies the padding and elevation in the random formations widget. Updates the HowToPlay widget by removing the text for the title and updating the layout for steps. Updates the text style for the title and description in the HowToPlayStep widget.